Add TS connector for user_logic
Adds TS connector for user_logic to insert users. At this point, it would probably be a good idea to also modify the schema and include a password field. Then, we can have a generateJWT function that will return the user's JWT if they already exist and their credentials are correct.